Авторизация и токен перенесены в контроллер пользователя
