+++ Knowtes.WebAPI/Controllers/UsersController.cs
@@ -2,6 +2,12 @@
 using CommonLibrary;
 using Knowtes.Backend.Models;
 using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Linq;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using Knowtes.WebAPI.AppData;
 
 namespace Knowtes.Backend.Controllers
 {
@@ -21,11 +27,59 @@ namespace Knowtes.Backend.Controllers
             return Ok(password);
         }
 
-        [HttpGet]
-        [Route("auth")]
-        public IActionResult Auth([FromQuery] string login, [FromQuery] string password)
+        // тестовые данные вместо использования базы данных
+        private List<User> people = new List<User>
         {
-            return Ok(new List<string> { login, password });
+            new User {email = "JellyBall@mail.ru", login = "eruwe", password = "qweasd123", name = "denis", Id = 1}
+        };
+
+        [HttpPost]
+        [Route("auth")]
+        public IActionResult Token([FromQuery] string username, [FromQuery] string password)
+        {
+            var identity = GetIdentity(username, password);
+            if (identity == null)
+            {
+                return BadRequest(new { errorText = "Invalid username or password." });
+            }
+
+            var now = DateTime.UtcNow;
+            // создаем JWT-токен
+            var jwt = new JwtSecurityToken(
+                    issuer: AuthOptions.ISSUER,
+                    audience: AuthOptions.AUDIENCE,
+                    notBefore: now,
+                    claims: identity.Claims,
+                    expires: now.Add(TimeSpan.FromMinutes(AuthOptions.LIFETIME)),
+                    signingCredentials: new SigningCredentials(AuthOptions.GetSymmetricSecurityKey(), SecurityAlgorithms.HmacSha256));
+            var encodedJwt = new JwtSecurityTokenHandler().WriteToken(jwt);
+
+            var response = new
+            {
+                access_token = encodedJwt,
+                username = identity.Name
+            };
+
+            return Ok(response);
+        }
+
+        private ClaimsIdentity GetIdentity(string username, string password)
+        {
+            User user = people.FirstOrDefault(x => x.login == username && x.password == password);
+            if (user != null)
+            {
+                var claims = new List<Claim>
+                {
+                    new Claim(ClaimsIdentity.DefaultNameClaimType, user.login)
+                };
+                ClaimsIdentity claimsIdentity =
+                new ClaimsIdentity(claims, "Token", ClaimsIdentity.DefaultNameClaimType,
+                    ClaimsIdentity.DefaultRoleClaimType);
+                return claimsIdentity;
+            }
+
+            // если пользователя не найдено
+            return null;
         }
     }
 }
